# feat: comprehensive Playwright E2E testing suite with dashboard fixes