--- a/.env.production.template
+++ b/.env.production.template
@@ -1,0 +1,175 @@
+# n8n-tweet Production Environment Configuration Template
+# このファイルを .env.production にコピーして実際の値を設定してください
+
+# =================================================================
+# 基本設定
+# =================================================================
+NODE_ENV=production
+LOG_LEVEL=info
+
+# ドメイン設定
+DOMAIN=yourdomain.com
+N8N_SUBDOMAIN=n8n
+DASHBOARD_SUBDOMAIN=dashboard
+
+# =================================================================
+# データベース設定
+# =================================================================
+DB_HOST=postgres
+DB_PORT=5432
+DB_NAME=n8n_production
+DB_USER=n8n
+DB_PASSWORD=your_secure_db_password_here
+
+# =================================================================
+# Redis設定
+# =================================================================
+REDIS_HOST=redis
+REDIS_PORT=6379
+REDIS_PASSWORD=your_secure_redis_password_here
+
+# =================================================================
+# n8n設定
+# =================================================================
+N8N_HOST=0.0.0.0
+N8N_PORT=5678
+N8N_PROTOCOL=https
+N8N_EDITOR_BASE_URL=https://n8n.yourdomain.com
+WEBHOOK_URL=https://n8n.yourdomain.com
+
+# n8n認証設定
+N8N_BASIC_AUTH_ACTIVE=true
+N8N_BASIC_AUTH_USER=admin
+N8N_BASIC_AUTH_PASSWORD=your_n8n_admin_password_here
+
+# =================================================================
+# ダッシュボード設定
+# =================================================================
+DASHBOARD_HOST=0.0.0.0
+DASHBOARD_PORT=3000
+
+# =================================================================
+# Twitter API設定
+# =================================================================
+# Twitter Developer Portalで取得したAPIキーを設定
+TWITTER_API_KEY=your_twitter_api_key_here
+TWITTER_API_SECRET=your_twitter_api_secret_here
+TWITTER_ACCESS_TOKEN=your_twitter_access_token_here
+TWITTER_ACCESS_TOKEN_SECRET=your_twitter_access_token_secret_here
+
+# Twitter API v2設定
+TWITTER_BEARER_TOKEN=your_twitter_bearer_token_here
+
+# =================================================================
+# セキュリティ設定
+# =================================================================
+# 暗号化キー（32文字のランダム文字列）
+ENCRYPTION_KEY=your_32_character_encryption_key_here
+
+# JWT秘密鍵（64文字のランダム文字列）
+JWT_SECRET=your_64_character_jwt_secret_here
+
+# セッション秘密鍵
+SESSION_SECRET=your_session_secret_here
+
+# =================================================================
+# CORS設定
+# =================================================================
+CORS_ORIGIN=https://dashboard.yourdomain.com
+
+# =================================================================
+# SSL/TLS設定
+# =================================================================
+SSL_CERT_PATH=/etc/letsencrypt/live/n8n.yourdomain.com/fullchain.pem
+SSL_KEY_PATH=/etc/letsencrypt/live/n8n.yourdomain.com/privkey.pem
+
+# =================================================================
+# 監視・ログ設定
+# =================================================================
+# Grafana設定（オプション）
+GRAFANA_ADMIN_USER=admin
+GRAFANA_ADMIN_PASSWORD=your_grafana_password_here
+
+# メトリクス設定
+ENABLE_METRICS=true
+METRICS_PORT=9090
+
+# =================================================================
+# メール設定（アラート用、オプション）
+# =================================================================
+SMTP_HOST=smtp.gmail.com
+SMTP_PORT=587
+SMTP_SECURE=false
+SMTP_USER=your_email@gmail.com
+SMTP_PASSWORD=your_email_password_here
+SMTP_FROM=no-reply@yourdomain.com
+
+# =================================================================
+# バックアップ設定
+# =================================================================
+BACKUP_ENABLED=true
+BACKUP_SCHEDULE=0 2 * * *  # 毎日午前2時
+BACKUP_RETENTION_DAYS=30
+BACKUP_S3_BUCKET=your-backup-bucket
+BACKUP_S3_ACCESS_KEY=your_s3_access_key
+BACKUP_S3_SECRET_KEY=your_s3_secret_key
+
+# =================================================================
+# パフォーマンス設定
+# =================================================================
+# ワーカープロセス数
+WORKER_PROCESSES=auto
+
+# メモリ制限
+NODE_OPTIONS=--max-old-space-size=2048
+
+# リクエスト制限
+RATE_LIMIT_WINDOW_MS=900000  # 15分
+RATE_LIMIT_MAX=100
+
+# =================================================================
+# RSS フィード設定
+# =================================================================
+# フィードの更新間隔（分）
+FEED_UPDATE_INTERVAL=30
+
+# AI関連度の最小スコア
+MIN_RELEVANCE_SCORE=0.5
+
+# 1日の最大ツイート数
+MAX_TWEETS_PER_DAY=10
+
+# =================================================================
+# タイムゾーン設定
+# =================================================================
+TZ=Asia/Tokyo
+GENERIC_TIMEZONE=Asia/Tokyo
+
+# =================================================================
+# デバッグ設定（開発時のみ）
+# =================================================================
+DEBUG=false
+ENABLE_DEBUG_LOGS=false
+
+# =================================================================
+# その他の設定
+# =================================================================
+# アプリケーション名
+APP_NAME=n8n-tweet
+APP_VERSION=1.0.0
+
+# サポート連絡先
+SUPPORT_EMAIL=support@yourdomain.com
+
+# 利用規約・プライバシーポリシー URL
+TERMS_URL=https://yourdomain.com/terms
+PRIVACY_URL=https://yourdomain.com/privacy
+
+# =================================================================
+# セキュリティ注意事項
+# =================================================================
+# 1. 本番環境では必ず強力なパスワードを設定してください
+# 2. 暗号化キーとJWT秘密鍵は必ずランダム生成してください
+# 3. このファイルは .gitignore に含まれていることを確認してください
+# 4. バックアップファイルにも機密情報が含まれることに注意してください
+# 5. 定期的にパスワードを変更してください