--- a/src/dashboard/index.js
+++ b/src/dashboard/index.js
@@ -117,7 +117,8 @@ class DashboardServer {
       contentSecurityPolicy: {
         directives: {
           defaultSrc: ["'self'"],
-          styleSrc: ["'self'", "'unsafe-inline'", 'https://cdn.jsdelivr.net', 'https://cdnjs.cloudflare.com'],
+          styleSrc: ["'self'", "'unsafe-inline'", 'https://cdn.jsdelivr.net',
+            'https://cdnjs.cloudflare.com'],
           scriptSrc: ["'self'", 'https://cdn.jsdelivr.net', 'https://cdnjs.cloudflare.com'],
           fontSrc: ["'self'", 'https://cdnjs.cloudflare.com'],
           imgSrc: ["'self'", 'data:', 'https:'],
@@ -140,6 +141,16 @@ class DashboardServer {
 
     // Static files middleware - serve dashboard UI
     const staticPath = path.join(process.cwd(), 'static')
+    this.logger.info('Static files path', { staticPath })
+
+    // Check if static directory exists
+    const fs = require('fs')
+    if (!fs.existsSync(staticPath)) {
+      this.logger.error('Static directory not found', { staticPath })
+    } else {
+      this.logger.info('Static directory found', { staticPath })
+    }
+
     this.app.use(express.static(staticPath, {
       index: 'index.html',
       maxAge: '1h',
@@ -210,6 +221,21 @@ class DashboardServer {
           description: 'Web-based monitoring dashboard'
         }
       })
+    })
+
+    // Root route - serve dashboard index.html
+    this.app.get('/', (req, res) => {
+      const fs = require('fs')
+      const indexPath = path.join(process.cwd(), 'static', 'index.html')
+
+      this.logger.info('Serving dashboard index', { indexPath })
+
+      if (fs.existsSync(indexPath)) {
+        res.sendFile(indexPath)
+      } else {
+        this.logger.error('Dashboard index.html not found', { indexPath })
+        res.status(404).send('Dashboard not found')
+      }
     })
   }
 
@@ -341,7 +367,7 @@ class DashboardServer {
       res.json({
         status: 'success',
         data: mockMetrics,
-        timeRange,
+        timeRange: req.query.timeRange || '1h',
         timestamp: new Date().toISOString()
       })
     }
@@ -424,7 +450,8 @@ class DashboardServer {
       const mockTweets = [
         {
           id: '1',
-          content: 'AI研究の最新論文: 「Transformer-based Language Models for Enhanced Natural Language Understanding」ArXivより #AI #研究 #NLP',
+          content: 'AI研究の最新論文: 「Transformer-based Language Models for ' +
+            'Enhanced Natural Language Understanding」ArXivより #AI #研究 #NLP',
           status: 'sent',
           category: 'ai',
           createdAt: new Date().toISOString(),
@@ -460,16 +487,16 @@ class DashboardServer {
       ]
 
       // Filter by status if specified
-      const filteredTweets = status
-        ? mockTweets.filter(tweet => tweet.status === status)
+      const filteredTweets = req.query.status
+        ? mockTweets.filter(tweet => tweet.status === req.query.status)
         : mockTweets
 
       res.json({
         status: 'success',
         data: filteredTweets,
         pagination: {
-          limit: options.limit,
-          offset: options.offset,
+          limit: parseInt(req.query.limit) || 100,
+          offset: parseInt(req.query.offset) || 0,
           total: filteredTweets.length
         },
         timestamp: new Date().toISOString()