… docker-compose.production.yml
… docs/deployment/proxmox-lxc-setup.md
… playwright-report/data/3623a9bc18a0168147dc7051d71abeafde67b099.md
… playwright-report/data/6ba0ffb017b99b45833f9106ba4d3808c1ed4b82.webm
… playwright-report/data/d606c7545cd71c2b0c27d9f99edbad4fa39f1174.png
… playwright-report/data/e139257baac1d69c8bffcb33219b65779cec0a26.png
… playwright-report/index.html
… scripts/deployment/quick-docker-setup.sh
… scripts/deployment/setup-production.sh
… src/middleware/security.js
… src/monitoring/metrics-collector.js
… src/utils/crypto.js
… src/utils/input-validator.js
… terraform/proxmox/main.tf
… tests/integration/comprehensive-e2e.test.js
… tests/integration/e2e-real-workflow.test.js
… tests/integration/end-to-end.test.js
… tests/integration/real-rss-test.js
… tests/integration/realistic-e2e.test.js
… tests/integration/tweet-generation-comprehensive-report.test.js
… tests/integration/tweet-generation-real-feed.test.js
… tests/integration/user-scenarios-e2e.test.js
… tests/manual/tweet-generation-demo.js
… tests/playwright/api.spec.js
… tests/playwright/dashboard.spec.js
… tests/playwright/performance.spec.js
… tests/playwright/screenshots/dashboard-desktop.png
… tests/playwright/screenshots/dashboard-main.png
… tests/playwright/screenshots/dashboard-mobile.png
… tests/playwright/screenshots/dashboard-navigation.png